fix: restructuring
